+++ examples/RSC/context/layout.storage.ts
@@ -1,0 +1,15 @@
+import { defineState, setupStorage } from "@yakocloud/state-vocab"
+
+export const layoutStorage = setupStorage({
+  session: defineState<{
+    id: number
+  }>({
+    defaultValue: {
+      id: -1
+    }
+  }),
+}, {
+  verbose: true,
+  verbosePath: "session",
+})
+