+++ examples/RSC/storage.client.ts
@@ -1,6 +1,0 @@
-"use client"
-
-import { storage } from "@/storage"
-import { clientify } from "@yakocloud/state-vocab/client"
-
-export const clientStorage = clientify(storage)