+++ examples/RSC/storage.server.ts
@@ -1,4 +1,0 @@
-import { storage } from "@/storage"
-import { serverify } from "@yakocloud/state-vocab/server"
-
-export const serverStorage = serverify(storage)